Move unit tests that consume job config outside of prow/.
